WEB_PLATFORM/requirements.txt
Django
requests
Pillow
psycopg2-binary
dj-database-url
django-redis
celery
python-consul2
